Change build directory
