+++ .gitignore
@@ -1,5 +1,8 @@
 node_modules/
 coverage/
-dist/
 npm-debug.log
 yarn-error.log
+
+# Autogenerated files
+index.js
+[a-z][a-z].js
