+++ .npmignore
@@ -11,3 +11,6 @@ yarn.lock
 .travis.yml
 .yaspellerrc
 coverage/
+tools/
+
+/[a-z][a-z].json
