+++ tools/templates/.eslintrc
