+++ tools/templates/locale.template.js
@@ -1,3 +1,4 @@
+/* eslint-disable quotes */
 module.exports = require('./convert')(
   __build.keysStr,
   __build.valuesStr
